packages/mgmt_server/pyproject.toml
[project]
name = "mgmt-server"
version = "0.1.0"
description = "Management API server for model/dataset management and task queue operations (port 9800)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "core",
    "cache",
    "database",
    "storage",
    "services",
    "fastapi[standard]>=0.128.0",
    "python-multipart>=0.0.0",
    "pwdlib[bcrypt]>=0.3.0",
    "bcrypt==4.1.2",
    "python-jose>=3.5.0",
    "cryptography>=42.0.0",
    "click>=8.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mgmt_server"]

[tool.uv.sources]
core = { workspace = true }
cache = { workspace = true }
database = { workspace = true }
storage = { workspace = true }
services = { workspace = true }
